requirements.txt
google-genai>=1.66,<2
pinecone>=8.1,<9
python-dotenv>=1.2,<2
sqlalchemy>=2.0,<3
psycopg[binary]>=3.3,<4
google-api-python-client>=2.192,<3
google-auth-oauthlib>=1.3,<2
